@@ -3,9 +3,9 @@ using Application.Dashboard.Queries.GetMenus;
 
 namespace WebAPI.Dashboard;
 
-public class DashboardModule : CarterModule
+public class DashboardModule : ICarterModule
 {
-    public override void AddRoutes(IEndpointRouteBuilder app)
+    public void AddRoutes(IEndpointRouteBuilder app)
     {
         var group = app.MapGroup("/api/dashboard");
 
