@@ -8,9 +8,9 @@ using Domain.Common;
 
 namespace WebAPI.Maintainer;
 
-public class InventoriesModule : CarterModule
+public class InventoriesModule : ICarterModule
 {
-    public override void AddRoutes(IEndpointRouteBuilder app)
+    public void AddRoutes(IEndpointRouteBuilder app)
     {
         var groups = app.MapGroup("api/maintainer/inventories")
             .RequireAuthorization(policy => policy.RequireRole("SuperAdmin", "Administrator"));
