@@ -4,7 +4,8 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 @Component({
     selector: 'app-dashboard',
     imports: [CommonModule],
-    template: './dashboard.component.html',
+    templateUrl: './dashboard.component.html',
+    styleUrl: './dashboard.component.scss',
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class DashboardComponent { }
